Update flag name for confirmation option and refactor confirmation logic
- Changed flag name from "confirm-before-write" to "confirm" in `cmd/root.go` for simplicity.
- Implemented a `confirm` function in `internal/process/write.go` to handle user confirmation input.
- Modified `WriteResult` function in `internal/process/write.go` to include an optional confirmation before writing to a file.
- Removed redundant `confirm` function in `internal/runner/runner.go` and used new confirmation workflow from `internal/process/write.go`.

+++ internal/process/write.go
@@ -1,11 +1,46 @@
 package process
 
 import (
+	"bufio"
 	"fmt"
+	"io"
 	"os"
+	"strings"
 )
 
-func WriteResult(outputText, outpath string) error {
+func confirm(message string) (bool, error) {
+	r := bufio.NewReader(os.Stdin)
+	for {
+		fmt.Printf("%s [y/N]: ", message)
+
+		res, err := r.ReadString('\n')
+		if err != nil {
+			if err == io.EOF {
+				return false, nil
+			}
+			return false, err
+		}
+		// Empty input (i.e. "\n")
+		if len(res) < 2 {
+			return false, nil
+		}
+
+		if strings.ToLower(strings.TrimSpace(res))[0] == 'y' {
+			return true, nil
+		}
+	}
+}
+
+func WriteResult(outputText, outpath string, needConfirm bool) error {
+	if needConfirm {
+		conf, err := confirm("May I write the results of this operation to a file?")
+		if err != nil {
+			return err
+		}
+		if !conf {
+			return nil
+		}
+	}
 	if err := os.WriteFile(outpath, []byte(outputText), 0644); err != nil {
 		return fmt.Errorf("error writing to file: %w", err)
 	}
